backend/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
pytz==2021.3
sqlparse==0.4.2
